requirements.txt
streamlit
pandas
plotly
psycopg2-binary
python-dotenv
openai
fastapi
uvicorn
pydantic
